pnet_packet: Move checksum benchmarks to benches directory
* These now work when running rustup run nightly cargo bench
