+++ pnet_packet/src/util.rs
@@ -236,21 +236,3 @@ mod tests {
         }
     }
 }
-
-#[cfg(test)]
-mod checksum_benchmarks {
-    use super::checksum;
-    use test::{black_box, Bencher};
-
-    #[bench]
-    fn bench_checksum_small(b: &mut Bencher) {
-        let data = vec![99u8; 20];
-        b.iter(|| checksum(black_box(&data), 5));
-    }
-
-    #[bench]
-    fn bench_checksum_large(b: &mut Bencher) {
-        let data = vec![123u8; 1024];
-        b.iter(|| checksum(black_box(&data), 5));
-    }
-}
